Add initial Emscripten scaffolding
Replaces initialization in main.c with main_js.c, which is simpler
(avoids most SDL initialization). Similarly screen.c is replaced
with screen_js.c, which sends the framebuffer contents to the JS
running in the Infinite Mac worker. jsgui.c and nd_js.cpp are stubs to
avoid other SDL dependencies.

This is enough to get the ROM monitor to be rendered, but no input or
disk I/O support yet.